chore(dev): [#64052] spread `pragma: no cover` et al
This commit spreads `pragma: no cover` and friends throughout the
entire codebase marking all the spots where the code is inadequately
tested (i.e. not even executed by the test-suite).

In the future all of these `pragma: no cover`s should be removed in
favor of testing of the code.