chore(dev): [#64052] add new no cover coverage comments
This commit introduces two new exclude comments to the coverage report
configuration. The two new exclude comments are:

* `# pragma: no cover file`
  Which excludes an entire file from the coverage report.

* `# coverage: pause` + `# coverage: unpause`
  Which temporarily pause and unpause coverage within a single file.
  This differs from the normal `# pragma: no cover` as it is not branch
  sensitive, which unfortunately is needed in some cases.

The implementation of these new comments is inspired by the multi-line
exclusion regexes examples on:

* https://coverage.readthedocs.io/en/latest/excluding.html

In the future all of these custom comments should be removed in favor of
testing of the code.