chore(config)!: [#63930] disallow overriding username generator objectClass
This commit makes overriding the `objectClass` within the `username_generator`
configuration of the `conversion_mapping` illegal. The variable must now
always be set to the default of `"UserNameGenerator"`.

BREAKING CHANGE:

It is no longer possible to override the `objectClass` configuration
parameter of the `username_generator`.

If you were setting it to `"AlleroedUserNameGenerator"` before, it must
now be removed and replaced with setting `remove_vowels` and
`disallow_mo_usernames` instead.