chore(config)!: [#63930] disallow setting username generator objectClass
This commit makes setting the `objectClass` within the `username_generator`
configuration of the `conversion_mapping` illegal. The variable now cannot
ever be set.

BREAKING CHANGE:

It is no longer possible to set the `objectClass` configuration
parameter of the `username_generator`.

If you were setting it to `"UserNameGenerator"` before, it must
now be removed, removing it changes nothing.