fix(ldap2mo): [#64329] remove employee speciality handling
This commit fixes an issue where attempting to write specific fields on
employees using the LDAP integration would overwrite unrelated fields.

The issue occurs because employee changes are handled in a special way
within the LDAP integration, such that they do not undergo the normal
codepath that implements patch write semantics.

The issue was introduced in version 1.0.0 of the integration and was
merged before very incremental change to the integration was released
with !101 (merged) (51e8391f).

This commit fixes the issue by removing the speciality employee handling
within the `SyncTool`'s `format_converted_objects` object, such that
`Employee` objects are not automatically short-circuited as creates, but
rather go through the normal flow, which consideres the current state of
the employee in MO just as is the case for all other object types.

The speciality handling code is however still needed, but it is now
contained within the `MOAPI`'s `edit_employee` function, which is now
called correctly whenever an employee is being updated.

This change ensures that employee objects have the expected patch
semantic as our generated changes are put atop the existing object
instead of overriding it entirely.

The solution still feels a little bit hacky as we should probably not be
calling `create_employee` from `edit_employee` and depending on an
undocumented feature / bug in MO to do so, but that is an issue for the
future.

The impact of the issue could be large as we may have overwritten
large amounts of important updates, unfortunately the only way to know
is to check the employee registration history for all employees for all
our customers and comparing each registration done by the LDAP
integration with the one prior keeping in mind that the configuration of
the LDAP integration may have changed over time.
This seems like a herculean task to undertake, so instead of taking on
Achelous we will simply bury our heads in the sand.