refactor(types): [#64329] spread the LDAPUUID type
This commit changes the codebase to utilize the LDAPUUID type wherever
possible to help distinguish which UUIDs in the program are LDAP entity
UUIDs and which are different kinds of MO UUIDs.

The refactoring is done on a best-effort basis throughout the codebase,
but there are likely minor mistakes and likely also spots that have been
missed.