test(types): [#64329] prove issues with NewType UUIDs
This commit proves the issues related to using `NewType` for making
specialized `UUID`s, namely that the `NewType` `UUID`s are not actually
`UUID`s at all, but rather just a static-only type-tag of whatever is
being put in.

This entirely breaks the assumption the codebase has had so far about the
use of `NewType`, and as such all the existing `NewType`s should
probably be revisisted at some point.