test(ldap2mo): [#64329] introduce test for MO reconciliation
This commit modifies the patch semantics test to affirm the state of drift
detection and reconciliation within the integration, i.e. affirming that
there currently is none.

The lack of drift detection and reconciliation means that changes to
fields owned by the integration are never overwritten by the integration,
and thus that data within MO cannot be trusted if multiple actors touch
the data.

The test is introduced without xfail to show that it passes as the code
is right now, i.e. that it passes without any drift detection and
reconciliation. It is hard however to test that something does *not*
happen eventually, so the test is essentially noop at this point, but
the commit is kept as it eases the reviewing process.