test(mo2ldap): [#64099] ensure create_user_trees LDAP account relevance
This commit modifies the test introduced in the prior commit such that
it now asserts the desired state of the code rather than the current
state, i.e. that existing LDAP accounts indeed are relevant, and that
`create_user_trees` should only be checked for `create`-operations.

The test must now be `xfail`ed as the desired state is not what is
exhibited by the code.