fix(mo2ldap): [#64099] ensure create_user_trees only affect creates
This commit fixes an issue where the `create_user_trees` check was being
performed on `edit`-operations rather than only on `create`-operations.

The issue was introduced in version 24.10.0 of the integration and was
merged before very incremental change to the integration was released
with !101 (51e8391f).

The issue was introduced in version 24.10.0 of the integration with !1108
(f6e1847a), which introduced the `create_user_trees` feature and
accidentally did not guard it to be applied for only `create`-operations.

This commit fixes the issue by delaying the `create_user_trees` checks
until we know whether the operation will be a `create` or not, and only
running the check iff it will be a `create`-operation.

The impact of the issue is likely quite small as we have only blocked
updates for LDAP accounts where the corresponding user now has an
engagement outside of our `create_user_trees`.
The workaround for this issue is to trigger a full resynchronization
once the integration has been rolled out.