test(mo2ldap): [#64099] prove create_user_trees LDAP account irrelevance
This commit adds a `create_user_trees` test to verify that the existence
of LDAP account(s) are irrelevance to the behavior of `create_user_trees`.

This is not the desired state of the code, as `create_user_trees` should
only apply to `create`-operations as given by the name, but the test is
simply introduced to prove the current behavior of the code, not the
desired behavior.

The test is introduced without xfail to show that it passes as the code
is right now, i.e. that it LDAP accounts are irrelevant.