feat(config): [#64131] add discriminator filter configuration variable
This commit introduces a new configuration variable `discriminator_filter`,
which takes a jinja template with the same format as the `discriminator_values`
and uses it to determine if a specific DN should be considered for the
discriminator itself.

This allows us to filter out certain accounts before discrimination,
which is useful in particular to entire ignore temporarily assigned
accounts for substitute or temporary employees.