fix(events): [#64435] fix an exponential number of issues
This commit fixes an issue where the number of messages in our queues
would grow exponentially (2^n) if both the reconcile and the ordinary
synchronization handlers would fail for the message message.

The issue arose from the fact that there are now two handlers on the
same routing-key, and the fact that each handler publishes the failing
message to its corresponding exchange instead of directly into its own
queue, i.e. if the reconciliation fails it will trigger both itself and
the ordinary synchronization, while if both fail the number of messages
will grow from 1 to 2 to 4 to 8 to 2^n.

The reason each handler republishes their message to the exchange is
that quorum queues in RabbitMQ does not correctly place rejected or
nacked messages to the back of the queue:

* https://github.com/rabbitmq/rabbitmq-website/discussions/1732

The issue was introduced for the MO2LDAP synchronization in version
30.2.0 of the integration with !1204 (eb8d9aac) and for the LDAP2MO
synchronization in version 30.1.0 of the integration with !1027 (710071e4).

The change is introduced without any tests whatsoever as we plan to
migrate the integration to MOs new event system as soon as possible.
Doing robust testing of this change therefore seems a waste of time.

The impact of the issue is that queues have been filled with a large
number of messages, however once this change is merged the queues should
eventually be processed and the issue should resolve itself.
There should be no impact on the data except a potential delay in
synchronization.