test(reconcile): [#64340] add test for the current ignore behavior
This commit adds an integration test for the `MO_UUIDS_TO_IGNORE` and
`LDAP_UUIDS_TO_IGNORE` to document the current behavior of those flags.

The current behavior of the code is not the desired one, however the
test is introduced to show the current behavior in favor of the desired
one to prove that the test works and passes.