feat(ldap2mo): [#64435] prevent unnecessary writes in MO
This commit ensures that noop synchronizations from LDAP to MO do not
result in new registrations in MO.

The code for this functional in version 1.0.0 of the integration and was
merged before very incremental change to the integration was released
with !101 (51e8391f), it was however disabled in version 2.7.3 of the
integration with !376 (87683a2b) and has remained disabled until now.

The code was commented out as a temporary measure till a better solution
to a specific problem and was unfortunately left that way until now.
Thus the integration have probably spammed MO with empty registrations
since then, and a registrations cleanup might be in order for all
customers.

This commit fixes the issue by simply reenabling the uncommented code,
as there seems to be no good reason for it to be disabled anymore
(nor a very good reason initially).