test(ldap2mo): [#64435] modify test for number registrations
This commit modifies our test for number of registrations given multiple
LDAP to MO writes from showing the current behavior of the code to
instead testing the desired behavior of the code, namely that noops do
not yield new registrations, i.e. that only actual changes do.

The test is now xfailed as the current behavior of the code is not the
desired state one.