fix(ldap2mo): [#64435] ensure mapping user-key achieves stable end-state
This commit fixes an issue where mapping user-key would trigger
resynchronizations forever as no stable end-state would be found.

The issue arose from the fact that we did not read out user-key from MO
when doing our comparison for changes, thus comparing the templated value
with the validator generated value (i.e. the UUID), and thereby trigger
resynchronizations forever.

The issue was introduced in version 1.0.0 of the integration and was
merged before very incremental change to the integration was released
with !101 (merged) (51e8391f).
It appears the issue was present in the very first version of the code
for reading employees.

This commit fixes the issue by simply reading the user-key from MO such
that our intermediate model has the fields that it requires.

The impact of the issue is that we may have been spamming MO with noop
writes since version 1.0.0, although we for sure have been spamming
since 2.7.3 due to the removal of code doing the comparison for changes.
Thus the impact of the issue itself is only that we have been spamming
MO with noop writes between version 1.0.0 and 2.7.3, from version 32.5.0
until now.