test(ldap2mo): [#64435] add test to show nickname mapping behavior
This commit adds a test to show that mapping nickname on employees is
broken, i.e. that trying to synchronize it does not actually bring MO to
a stable end-state, thus triggering resynchronizations forever via
reconciliation.

Additionally the test shows that the issue is only present with empty
values, i.e. that actually synchronizing a nickname is stable.