test(ldap2mo): [#64435] modify test for nickname mappings
This commit modifies our test for nickname mapping from showing the
current behavior of the code to instead testing the desired behavior
of the code, namely that nickname mappings can be synchronized and
result in stable end-state.

The test is now xfailed as the current behavior of the code is not the
desired state one.