fix(ldap2mo): [#64435] ensure mapping nickname achieves stable end-state
This commit fixes an issue where mapping nickname would trigger
resynchronizations forever if set to an empty value as no stable
end-state would be found.

The issue arose from a combination of multiple factors.
First up, the LDAP integration throws away falsy values from the
template result, thus leaving us the default value from the intermediate
model, thereby breaking all empty value comparisons (i.e. `null` vs `""`).
Secondly, MO states that the nickname fields are nullable, but seems to
return `""` instead of `null` whenever they are empty.

This means that even if the LDAP integration attempted to template out
the expected value from MO (i.e. `""`), it would not compare correctly,
because the templated value would be falsy and thus removed and replaced
with the model default (i.e. `null`) breaking the comparison.

However one could also argue that the core issue that we write `null` as
nickname to MO due to the above, but end up reading out `""` instead,
thus correctly necessitating another write.

The issue was introduced in version 1.0.0 of the integration and was
merged before very incremental change to the integration was released
with !101 (51e8391f). It appears however that the defective
check was introduced !9 (26a844f3).

This commit implements a dirty hotfix that simply changes the default on
the intermediate model to fit MOs incorrect response, thereby effectively
not fixing any of the underlying issues, but merely resolving the symptom
observed from them.

The impact of the issue is that we may have been spamming MO with noop
writes since version 1.0.0.