fix(reconcile): [#64435] reject deleted LDAP UUIDs
This commit fixes an issue where LDAP UUIDs could get stuck forever in
the LDAP reconciliation queue, if their corresponding account in LDAP
disappeared before the UUID was processed.

The message would get stuck as we would continuously try to lookup the
UUID and fail repeatedly in an infinite loop.

The issue was introduced in version 30.2.0 of the integration with !1207
(710071e4), which introduced the LDAP reconciliation logic to the
integration, but forgot to take into account this error mode.

This commit fixes the issue by copying the equivalent code from the
normal LDAP UUID handler, i.e. by rejected the deleted UUIDs.

In the future we should probably handle these situations to delete data
in MO, using some sort of state to map the LDAP UUID to a corresponding
MO UUID.

The impact of the issue is minor as messages have still been processed,
although at a reduced rate.