test(usernames): [#64038] modify test to desired state
This commit modifies the test introduced in the prior commit such that
it describes the desired state of the integration rather than the
current state of the integration, i.e. that usernames are only tested a
single time for correctness instead of 8 times.

The test is also marked as xfailed to signify that the desired behavior
of the integration is not the current behavior.