feat(usernames): [#64131] add support for reusing usernames
This commit adds support for reusing usernames adding the new
configuration variable `REUSE_OLD_USERNAMES` to the `username_generator`
section of the configuration.

The new configuration variable is disabled by default, but can be
enabled to get the new behavior.

The newly introduced unittest is also adjusted such that the new
configuration variable is set, and the xfail on the desired state is
removed verifying that the implementation works as expected.