refactor(mo2ldap): [#64038] introduce NoGoodLDAPAccountFound exception
This commit introduces the `NoGoodLDAPAccountFound` exception to the
`_find_best_dn` method to be raised in the case that all found LDAP
accounts have been filtered away using the discriminator.

It is introduced to try to simplify the return-value of the
`_find_best_dn` method, such that it is explicit that this situation was
encounted, instead of having this information encoded as a missing DN
along with a `create=False` flag.

The refactoring is done in preparation for splitting `_find_best_dn` in
two.