docs(mo2ldap): [#64038] move make_mo_employee_dn comments to account creation
This commit moves and reworks bunch of comments from the
`make_mo_employee_dn` method to the code creation employees, where the
comments are actually needed.

The prior placement was a leftover from when generating DNs had the side
effect of also creating a partly done account in LDAP, but the code has
since been refactored such that the entire account is being built
atomically in fully in one place.