fix(mo2ldap): [#64038] use templated common name (if any) for DN
This commit fixes an issue where common-name mappings in the jinja
template output would be entirely ignored during creates instead being
replaced with an automatically generated common-name.

The issue arose from a combination of two factors, firstly the logic used
to generate DNs which explicitly set the common-name to the value
returned by the `generate_common_name` method, entirely ignoring whatever
was templated out for the common name.
Secondly from the fact that writes to both DN and CN at once favors the
value written to the DN, thus overruling the value set in the CN.

This means that even if the LDAP integration was configured to template
out a very specific naming scheme, an entirely different naming scheme
would be utilized for creates, only to be corrected later when an edit
to the same entity occurs, or alternatively only to block all edits to
the object as the edit would trigger a name conflict, thus leaving the
object in an incorrect state forever.

The issue may have been hidden by the fact that writes to LDAP used to
be on a best-effort approach, with errors being silently logged and
otherwise ignored, thus effectively hiding the issue.
Additionally it was not until version 20.1.1 of the integration with
!1008 (78709cea) that it even became possible to write attributes
contained in the DN.

The issue can be attributed to many different versions, but the earliest
version of the issue was introduced in version 1.0.0 of the integration
and was merged before very incremental change to the integration was
released with !101 (51e8391f).
It appears however that the defective pattern of added users without
their corresponding data, and thus the broken multistage creation of
users was introduced in !3 (bc53cfff).

This commit implements a fix by utilizing the common-name from the jinja
template output when generating the DN for newly created users, ensuring
that the common-name templated and provided as an attribute fits with
the common-name utilized in the generated DN.

The impact of the issue is that we may have been generating users with
usernames not adhering to the naming guidelines since version 1.0.0, and
that users may have gotten stuck in this state due to renaming blocking
the edits, although prior versions of the LDAP integration would not
have this issues as they would simply write objects partly leaving them
in an undefined state where some fields would be up-to-date and others
would be left behind.