test(mo2ldap): [#64038] confirm current common name templating behavior
This commit adds a test to show the current behavior of templating
common-names, i.e. that they are entirely ignored during creates and only take
effect during edits.

This behavior is due to how DNs are generated separately from the entire
desired ldap state as produced by the template engine.