fix(usernames): [#64951] filter out superfluous search result references
This commit fixes an issue where the username generator would reject all
username candidates when running under Microsoft's Active Directory due
to superfluous search result references not being handled by the ldap3
library.

The issue arose from the expectation that the ldap3 library would be
able to resolve search result references by issuing sub-search queries
on our behalf under Active Directory as it does when running against a
Standard LDAP server, however it seems like ldap3 cannot handle all of
Active Directory's shenanigans and occasionally just give up leaving the
search result references in the search response.

This meant that all generated username candidates were rejected by the
fact that the search response contained leftover search result references.

The issue was introduced in version 32.14.2 of the integration when a
performance optimization was made to the code checking for used LDAP
usernames. The change was made in !1274 (612b5ca8) and modified the
username checking from reading all usernames in all of LDAP to instead
checking a specific username using a search operation.

This commit "fixes" the issue by assuming that the search result
references left unresolved by the ldap3 library are superfluous and as
such that it is okay to simply ignore them.
This assumption may be proven wrong in the future, but for now we assume
that it is okay.

The impact of the issue is that we may have been unable to generate
usernames in the versions since 32.14.2, and thus that user creation
have been blocked, leaving messages stuck in the queue.
Deploying this fix should allow all of these messages to pass by restoring
the integrations ability to generate usernames.