feat(mo2ldap): [#64963] introduce support for mapping any object class
This commit introduces the necessary code to implement listeners and
handlers for the new `mo_to_ldap` configuration mapping which was
introduced in a recent commit.

The listeners utilize MOs new GraphQL based-event system via the
FastRAMQPI bridge-service which calls HTTP endpoints with the fetched
GraphQL events.
Thus to configure the new setup a listener is configured for each
mapping attaching to the configured routing-key in MO, mapping it to a
generated URL path on the integration itself, at which a router is
mounted using the same mapping to expose an endpoint handler partialed
with the template and such from the configuration mapping.

As such an event on the GraphQL event system will trigger a call to the
generated HTTP endpoint and thus trigger the partialed handler to
process the incoming event in a generic way.

The generic handler extracts the UUID from the incoming MO event before
passing it into the configured Jinja template expecting it to generate a
JSON output with the `dn`, `create` and `attributes` keys, which it then
passes to the `ensure_ldap_object` method on the `LDAPAPI` to ensure
that the templated state is reached in LDAP.