feat(mo2ldap): [#64963] introduce ensure_ldap_object method
This commit introduces the `ensure_ldap_object` method which allows a
single entrypoint to both create and modify an ldap object. Whether the
create or modify logic is used is determined by the `create` flag.

The code for the update stage of the code is heavily inspired by the
contents of the `listen_to_changes_in_employees` method on the `SyncTool`
class, which it will replace in a future commit.

The code is extracted as it will be used from multiple handlers in the
future.