feat(mo2ldap): [#64963] convert create_ituser_link into ensure_ituser_link
This commit changes the `create_ituser_link` create method into an
ensure method called `ensure_ituser_link` instead.

The method now checks if the ituser link already exists, and if it does,
does nothing, only creating the link iff it does not already exist.

The change is made in preparation of the introduction of an ensure LDAP
object method, which will serve to eliminate the separation of the
create and update flows of the code, thus necessitating that this method
has ensure semantics as well.