test(mo2ldap): [#65019] modify multi-valued RDN renaming behavior test
This commit modifies the test introduced in the prior commit to test for
the desired behavior of multi-valued RDN renaming instead of simply
showing the current behavior.

The desire behavior is that we do not destroy the existing multi-valued
nature of the RDN to only keep the latest updated field, but rather keep
the format and only change the values within it.

This change requires marking the test as xfailed since the current
behavior is not the desired behavior.