chore(ldap): [#65020] add support for RFC4528 Assertion Control
This commit adds support for RFC4528 Assertion Control, which can be
used to ensure that a write operation only takes place if a provided
search filter matches an entry on the server.

It is similar in purpose to ETag (entity tag) within HTTP.

This commit only adds a function to construct an assertion control tuple
along with expanding the `ldap_modify` to pass through controls to the
`ldap3` library. It does not actually integration assertion control with
the rest of the integration, this is left as a task for a future commit.

Additionally a couple of integration tests have been added to prove that
assertion controls indeed work, and are ready to be used in the integration.