chore(ldap): [#65020] add assertion filter generator
This commit adds a helper function `construct_assertion_control_filter`
to generate RFC4515 filters for use in RFC4528 Assertion Control
assertions.

The output of the function can be used as input to the
`construct_assertion_control` function to create an assertion control
3-tuple including the BER-encoded filter.

This commit only adds the generator function without actually integrating
it into the code-base itself. The task of doing so is left for a future
commit.