chore(ldap): [#65020] add support for empty list assertions
This commit adds support for empty-lists as values in the attributes
dictionary to `construct_assertion_control_filter`.

The filter implementation for this check is a little convoluted, since
LDAP does not support filtering for empty values forcing us to rather
look for the inverted wildcard match (i.e. not anything matches).