chore(ldap): [#65020] add support for space-only strings
This commit adds support for matching space-only attribute values in
`construct_assertion_control_filter` while also ensuring that spaces are
handled unambiguously in general to avoid misinterpretation by the
domain controller.

While RFC4515 is explicit about how to handle certain specialty
characters, such as `*`, `()` and `\`, it does not explicitly state how
to handle spaces leaving it up to the LDAP implementation to decide how
these should be handled.

LDAP implementations are generally well-behaved, but may in some cases
attempt to strip extraneous whitespace in the filter string, which may
result in a filter such as `cn= ` becomes `cn=` thus changing the
desired semantics of the filter.
This can be avoided by intentionally octet-encoding the space as `"\20"`
instead of simply `" "` which seems to be the de-facto way to
communicate to the domain controller that the space is a desired
character, rather than strippable whitespace.

It is unfortunate that the behavior is not standardized, but using
octet-encoding seems to work and achieve the desired effect.