feat(ldap): [#65020] introduce assertion control in modify_ldap_object
This commit introduced preliminary support for assert controls within
the `modify_ldap_object` method, allowing calls to specify an expected
before-state (or old-state) that be present when the write is attempted.

The goal of the change is to allow conditional writes such that multiple
writers will not accidentally override each others data resulting in
datalosses.

While this is an unlikely event to occur for employees due to the low
change frequency and the fact that the integration locks itself out of
simultaneous processing of the same event UUIDs, however with the coming
introduction of shared groups having multiple entities in MO attempting
to modify the same LDAP object this should come in handy to ensure a
paradigm of optimistic locking can be upheld.