chore(ldap2mo)!: [#59372] remove support for mapping lists to multiple objects
This commit adds a check which refuses synchronization if any of the
mapped attributes has a higher cardinality than 1, i.e. lists attributes
with more than 1 element.

The functionality is removed as the code implementing it was highly
confusing and as it does not seem to be have been used in any environment.

BREAKING CHANGE:

It is no longer possible to map list attributes with length larger than 1.
There is no workaround.