refactor(ldap2mo): [#59372] eliminate list wrapping on from_ldap return type
This commit removes the list wrapping on the return type of the
`from_ldap` method such that the constructed objects are returned
directly instead of being returned as a single element list.

The list construction is not eliminated entirely, but rather has been
moved into the calling function, from which a future commit can
eliminate it entirely.