chore(mo2ldap): [#59372] separate termination as an early stage
This commit separates the handling of termination within the `from_ldap`
method into its own separate stage minimizing its coupling with the
remainder of the method.

While the code becomes longer, the change allows us to now exclude all
custom fields on the `LDAP2MOMapping` in the `get_fields` helper, such
that the remainder of the `from_ldap` method does not have to deal with
speciality fields, but rather can process only MO mapping fields.