feat(mo2ldap): [#64963] added ituser_uuid_to_person_uuid filter
This commit introduces the `ituser_uuid_to_person_uuid` filter function
to the Jinja context which as the name implies help to convert an ituser
UUID to a corresponding person UUID (if any).

In the future when we start optimizing the performance of the
integration we may want to simply fetch the person UUID directly when
fetching the itusers or alternatively cache all our reads within a
single event-handling, but for now we accept that the endpoint does not
feature good performance nor make the user consider how to achieve good
performance.