feat(mo2ldap): [#64963] expose get_person_dn as filter
This commit exposes the `get_person_dn` function as a jinja filter, such
that it can be used as a filter within our Jinja templates, enabling
its use in Jinja `map`s using a syntax alike: `persons|map('get_person_dn)`

The filter translates a MO person UUID to a corresponding DN.
For details see the commit which originally introduced it as a global.

The commit also lies the foundation for future filters to be added in a
similar way such that future functions can be introduced either as
filters or as globals depending on what the implementer finds preferable.