chore(ldap2mo): [#64003] missing values in LDAP is now an error
This commit makes missing values in LDAP an error by logging to the
error log instead of the info log.

This makes sense because we assume that values are always present, and
if this is not the case we expect the user to wrap they possibly empty
variables in `skip_if_none`.