test(ldap2mo): [#64963] remove unnecessary asyncio.sleep mocking
This commit removes an unnecessary `asyncio.sleep` mock statement
which replaced the sleep with an `AsyncMock`.

The mock statement caused issues after converting the test-suite to
using uvicorn instead of asgi-lifespan as the integration-test were
killed due to an out-of-memory (OOM) error when running under CI, itself
due to a "memory-leak" within the integration-tests.

The issue itself was flaky and thus hard to pin down, and while it has
not been pinned down entirely, a good explaination has been found.

The issue was caused by the `_setup_periodic_metrics` function inside af
FastRAMQPI, specifically inside of RAMQP. The function setups Prometheus
metrics and refreshes them on a schedule using `asyncio.sleep` to delay
the periodic updates, thus the mocking of `asyncio.sleep` has caused two
effects;

1. The function would not be sleeping, and thus would run as fast as the
   CPU can manage.
2. Every single call to `asyncio.sleep` would be registered on the mock,
   such that it could be checked later using `mock.calls`.

This together meant that the `_setup_periodic_metrics` would quickly
consume vast amounts of memory, not exactly memory-leaking, but simply
using it for an unwanted purpose.

Additionally a problem was found with the cancellation of
`_setup_periodic_metrics` within FastRAMQPI which caused the task to
never get cancelled and thus made the issue run forever, thus allowing
the "memory-leak" to reach the point where it would trigger an
out-of-memory situation.

The FastRAMQPI cancellation issue has been resolved by:

* https://git.magenta.dk/rammearkitektur/FastRAMQPI/-/merge_requests/378

It is however still unknown why the change from asgi-lifespan to uvicorn
uncovered these issues.

The solution to the issue however was a simple one; simply removing the
mocking of `asyncio.sleep` as it is no longer required.
It was originally introduced to avoid the test waiting for 30 seconds
for the quorum queue workaround to republish messages to ensure cyclic
message behavior, however since then the HTTP handler has been converted
to not cycle messages with this behavior now being isolated to the AMQP
handler instead.